fixed bug in pytorch data teacher
